+++ src/main/java/view/graphical/FilterByCategory.java
@@ -1,6 +1,7 @@
 package view.graphical;
 
 import controller.data.YaDataManager;
+import controller.share.MenuHandler;
 import javafx.event.ActionEvent;
 import javafx.scene.control.Alert;
 import javafx.scene.control.TextField;
@@ -14,9 +15,14 @@ public class FilterByCategory {
         String categoryName = categoryNameTextField.getText();
         try {
             View.filteringMenu.filter(new model.filter.FilterByCategory("Filter by category "+categoryName , View.manageCategoryMenu.getCategory(categoryName)));
-            SceneHandler.getProductsMenuLoad().deleteCommodities(SceneHandler.getProductsMenuLoad().getRoot());
-            SceneHandler.getProductsMenuLoad().setCommodities(SceneHandler.getProductsMenuLoad().getRoot());
-
+            if (MenuHandler.getInstance().getCurrentMenu().getFxmlFileAddress().equals("../../Products.fxml")) {
+                SceneHandler.getProductsMenuLoad().deleteCommodities(SceneHandler.getProductsMenuLoad().getRoot());
+                SceneHandler.getProductsMenuLoad().setCommodities(SceneHandler.getProductsMenuLoad().getRoot());
+            }
+            if (MenuHandler.getInstance().getCurrentMenu().getFxmlFileAddress().equals("../../fxml/OffMenu.fxml")) {
+                SceneHandler.getOffMenu().deleteCommodities(SceneHandler.getOffMenu().getRoot());
+                SceneHandler.getOffMenu().setCommodities(SceneHandler.getOffMenu().getRoot());
+            }
         } catch (Exception exception) {
             Alert alert = new Alert(Alert.AlertType.ERROR);
             alert.setTitle("Invalid category name");
